+++ src/main/java/com/example/demo/hello/domain/user/Form/HidukeOutForm.java
@@ -1,4 +1,4 @@
-package com.example.demo.hello.domain.user.entity;
+package com.example.demo.hello.domain.user.Form;
 
 import java.time.LocalDate;
